Pr 1958 hero img flicker (#5788)
* fix: remove replace hydration method

* fix: remove replace hydration method

* fix: removed unnecessary breakpoint condition from nav template

* fix: minified error in payment nav

* fix: minified error in layout overlay

* fix: minified error in payment methods page

* fix: minified error in layout overlay

* fix: minified error in use visible content

* fix: minified error in trade types

* fix: minified error in layout overlay

* fix: minified error in cfd page

* fix: minified error in market pages

* fix: minified error in regulatory page

* fix: minified error in forex market page

* fix: minified error in synthetic market page

* fix: minified error in rest market pages

* fix: remove unused style from deriv prime

* fix: minified error in deriv prime

* fix: minified error in affiliate ib's hero and numbers component

* fix: minified error in affiliate ib's why us component

* fix: minified error in affiliate ib page

* fix: minified error in help centre page

* fix: minified error in etf page

* fix: minified error in help centre page

* fix: minified error in market pages

* fix: added fallback to prime suspense

* fix: affiliate ib

* fix: signup form
