+++ crowdin/messages.json
@@ -535,6 +535,7 @@
   "517286342": "<p>To place a vanilla options contract, you’ll need to:</p><ul><li>Select the underlying asset you want to trade.</li><li>Determine your contract duration or end time.</li><li>Predict the asset’s price movement by choosing Call or Put.</li><li>Select your preferred strike price</li><li>Enter your stake.</li><li>Purchase the contract</li></ul>",
   "517833647": "Volatility 50 (1s) Index",
   "517844699": "Alibaba Group Holding Ltd",
+  "518861708": "Open a demo Deriv MT5 or demo Deriv X account and practise with an unlimited amount of virtual funds.",
   "518906831": "Can I use the same address to make a cryptocurrency deposit?",
   "519988278": "Fox Corp Class B",
   "520136698": "Boom 500 Index",
@@ -4110,6 +4111,9 @@
   "-2121510827": "Sweden",
   "-1357648282": "For fair resolution of any complaints, please <0>chat</0> with us. To learn more, see our <1>complaint policy</1>.",
   "-1838458999": "For fair resolution of any complaints, please <0>chat</0> with us.",
+  "-1131400885": "Deriv Investments (Europe) Limited",
+  "-1231248751": "Deriv Investments (Europe) Limited (Company No. C 70156), incorporated on the 22nd April 2015, is registered in Malta with its registered office located at Level 3, W Business Centre, Triq Dun Karm, Birkirkara BKR9033, Malta. Deriv Investments (Europe) Ltd is licensed in Malta and regulated by the Malta Financial Services Authority under the Investments Services Act <0>(view licence)</0> to provide investment services.",
+  "-2051445931": "Clients in the European Union who wish to trade financial instruments can have accounts under Deriv Investments (Europe) Limited.",
   "-1446148261": "CFDs - Forex",
   "-1121949345": "CFDs - Stock indices",
   "-33096025": "CFDs - Cryptocurrencies",
@@ -4117,9 +4121,6 @@
   "-1238176624": "CFDs - Synthetics: Volatility 250 (1s) Index",
   "-1076679821": "Multipliers - Cryptocurrencies",
   "-812435767": "Multipliers - Synthetics: Crash 300 Index",
-  "-1131400885": "Deriv Investments (Europe) Limited",
-  "-1231248751": "Deriv Investments (Europe) Limited (Company No. C 70156), incorporated on the 22nd April 2015, is registered in Malta with its registered office located at Level 3, W Business Centre, Triq Dun Karm, Birkirkara BKR9033, Malta. Deriv Investments (Europe) Ltd is licensed in Malta and regulated by the Malta Financial Services Authority under the Investments Services Act <0>(view licence)</0> to provide investment services.",
-  "-2051445931": "Clients in the European Union who wish to trade financial instruments can have accounts under Deriv Investments (Europe) Limited.",
   "-112814932": "Deriv (FX) Ltd",
   "-1269257733": "Clients in the rest of the world (except for restricted countries such as the USA, Canada, and Hong Kong) who wish to trade CFDs on forex and cryptocurrencies can have MT5 Financial STP accounts under Deriv (FX) Ltd.",
   "-2019617323": "Deriv (BVI) Ltd",
@@ -4209,9 +4210,6 @@
   "-943253121": "The term ‘margin’ refers to the deposit needed to open a leveraged position, which is a position larger than your capital investment and leads to increased market exposure.",
   "-339497727": "Trade types margin market",
   "-1009744297": "Use our margin calculator to calculate the margin required to increase your market exposure (the market value of your position) on Deriv’s CFD trading platforms.",
-  "-1603400333": "Open a demo CFDs account and practise with an unlimited amount of virtual funds.",
-  "-446423672": "Open a demo Deriv MT5, Deriv X, or Deriv cTrader account and practise with an unlimited amount of virtual funds.",
-  "-1978157334": "Trade with a real Deriv MT5, Deriv X, or Deriv cTrader account and get access to high leverage to trade positions larger than your existing capital.",
   "-1207172200": "Start trading CFDs on Deriv",
   "-1703690406": "When you trade CFDs on margin, you increase your market exposure, thus amplifying both your potential profit and potential loss.",
   "-843831637": "Stop loss",
@@ -4248,6 +4246,8 @@
   "-894512007": "favourite markets icon",
   "-252740007": "instant access icon",
   "-1156092900": "flexible icon",
+  "-1603400333": "Open a demo CFDs account and practise with an unlimited amount of virtual funds.",
+  "-1242544018": "Trade with a real Deriv MT5 or real Deriv X account and get access to high leverage to trade positions larger than your existing capital.",
   "-544150655": "Markets available for margin trading",
   "-867909844": "Access over 30 assets and also trade with up to 1:30 leverage and zero commission on trades.",
   "-309438223": "Trade synthetic indices on margin 24/7 without being affected by natural events and liquidity risks.",
