+++ src/features/pages/partners/deriv-prime/connectivity-reliability/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { connectivity_items } from './data'
 import ConnectivityReliabilityItem from './connectivity-reliability-item'
-import { item_container } from './styles.module.scss'
 import Container from 'features/components/atoms/container'
 import Typography from 'features/components/atoms/typography'
 import { Localize } from 'components/localization'
