+++ src/pages/markets/components/sections/_nav-tab.tsx
@@ -7,6 +7,7 @@ import useRegion from 'components/hooks/use-region'
 import device from 'themes/device'
 import { usePlatformQueryParam } from 'components/hooks/use-platform-query-param'
 import { TNavTab, TTabButton, TTabList } from 'pages/markets/static/content/_types'
+import { usePageLoaded } from 'components/hooks/use-page-loaded'
 
 const TabsContainer = styled(Flex)`
     justify-content: center;
