+++ src/pages/markets/stock/index.tsx
@@ -7,18 +7,25 @@ import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
 import { SEO } from 'components/containers'
 import { TGatsbyHead } from 'features/types'
+import { usePageLoaded } from 'components/hooks/use-page-loaded'
 
-const StocksAndIndicesPage = () => (
-    <Layout type="noNav">
-        <DerivedFXHero
-            title="_t_Stocks & indices_t_"
-            description="_t_Trade global stocks of your favourite household brands and international stock market indices on Deriv. Expand your trading opportunities with access to a wide range of stocks and indices at competitive prices, and the convenience to trade indices beyond standard market hours._t_"
-            is_stocks_and_indices
-        />
-        <StockIndices simple_step_content={simple_step_content_stock} />
-        <SignupPublic />
-    </Layout>
-)
+const StocksAndIndicesPage = () => {
+    const [is_mounted] = usePageLoaded()
+
+    if (!is_mounted) return null
+
+    return (
+        <Layout type="noNav">
+            <DerivedFXHero
+                title="_t_Stocks & indices_t_"
+                description="_t_Trade global stocks of your favourite household brands and international stock market indices on Deriv. Expand your trading opportunities with access to a wide range of stocks and indices at competitive prices, and the convenience to trade indices beyond standard market hours._t_"
+                is_stocks_and_indices
+            />
+            <StockIndices simple_step_content={simple_step_content_stock} />
+            <SignupPublic />
+        </Layout>
+    )
+}
 
 export default WithIntl()(StocksAndIndicesPage)
