+++ src/pages/regulatory/_type.ts
@@ -1,0 +1,3 @@
+export type RegulatoryProps = {
+    language: string
+}
