+++ src/pages/trade-types/content/static/_cfd.tsx
@@ -4,7 +4,7 @@ import Favourite from 'images/svg/markets/forex/favourite.svg'
 import Golong from 'images/svg/markets/forex/go-long.svg'
 import Instant from 'images/svg/markets/forex/instant.svg'
 import SwapFree from 'images/svg/markets/forex/swap-free.svg'
-import { TRegion, TString } from 'types/generics'
+import { TRegion, TSmartContent, TString } from 'types/generics'
 
 type TCFD = {
     text: TString | ((config: Partial<TRegion>) => TString)
@@ -15,40 +15,84 @@ type TCFD = {
     alt: TString
 }
 
-export const cfd_content: TCFD[] = [
+type CFDConfig = {
+    is_eu: boolean
+}
+
+type SmartCFD = TSmartContent<TCFD, CFDConfig>
+
+export const cfd_content: SmartCFD[] = [
     {
-        title: ({ is_eu }) =>
-            is_eu ? '_t_Tight spreads_t_' : '_t_High leverage, tight spreads_t_',
-        text: ({ is_eu }) =>
-            is_eu
-                ? '_t_Take advantage of tight spreads on Deriv’s CFD trading platforms._t_'
-                : '_t_Take advantage of high leverage and tight spreads on Deriv’s CFD trading platforms._t_',
-        src: Tight,
-        alt: '_t_High leverage, tight spreads_t_',
+        id: 0,
+        data: {
+            title: '_t_High leverage, tight spreads_t_',
+            text: '_t_Take advantage of high leverage and tight spreads on Deriv’s CFD trading platforms._t_',
+            src: Tight,
+            alt: '_t_High leverage, tight spreads_t_',
+        },
+        visibility: {
+            is_eu: false,
+        },
     },
     {
-        title: '_t_All your favourite markets_t_',
-        text: '_t_Trade on financial markets plus our proprietary synthetics that are available 24/7._t_',
-        src: Favourite,
-        alt: '_t_All your favourite markets_t_',
+        id: 11,
+        data: {
+            title: '_t_Tight spreads_t_',
+            text: '_t_Take advantage of tight spreads on Deriv’s CFD trading platforms._t_',
+            src: Tight,
+            alt: '_t_High leverage, tight spreads_t_',
+        },
+        visibility: {
+            is_eu: true,
+        },
     },
     {
-        title: '_t_Go long or short_t_',
-        text: '_t_Open long or short positions based on your predictions of the market price movements._t_',
-        src: Golong,
-        alt: '_t_Easy deposits and withdrawals_t_',
+        id: 1,
+        data: {
+            title: '_t_All your favourite markets_t_',
+            text: '_t_Trade on financial markets plus our proprietary synthetics that are available 24/7._t_',
+            src: Favourite,
+            alt: '_t_All your favourite markets_t_',
+        },
     },
     {
-        title: '_t_Expert and friendly support_t_',
-        text: '_t_Get expert, friendly support when you need it._t_',
-        src: FriendlySupport,
-        alt: '_t_Expert and friendly support_t_',
+        id: 2,
+        data: {
+            title: '_t_Go long or short_t_',
+            text: '_t_Open long or short positions based on your predictions of the market price movements._t_',
+            src: Golong,
+            alt: '_t_Easy deposits and withdrawals_t_',
+        },
     },
     {
-        title: '_t_Instant access_t_',
-        text: '_t_Open an account and start trading in minutes._t_',
-        src: Instant,
-        alt: '_t_Instant access_t_',
+        id: 3,
+        data: {
+            title: '_t_Expert and friendly support_t_',
+            text: '_t_Get expert, friendly support when you need it._t_',
+            src: FriendlySupport,
+            alt: '_t_Expert and friendly support_t_',
+        },
+    },
+    {
+        id: 4,
+        data: {
+            title: '_t_Instant access_t_',
+            text: '_t_Open an account and start trading in minutes._t_',
+            src: Instant,
+            alt: '_t_Instant access_t_',
+        },
+    },
+    {
+        id: 5,
+        data: {
+            title: '_t_Swap-free MT5 account_t_',
+            text: '_t_Leave your trades open overnight with no additional charges._t_',
+            src: SwapFree,
+            alt: '_t_Swap free_t_',
+        },
+        visibility: {
+            is_eu: false,
+        },
     },
 ]
